system: extend rtc implementation to really wakeup an android device
On Android just setting a RTC alarm doesn't wakeup the device when in deep sleep. We have
to set a second alarm timer through a Android specific kernel interface (/dev/alarm) in
order to make sure the device wakes up when we need it. RTC applies on top of it so if we
set a RTC alarm for 10 seconds in the future and configure the Android alarm timer the
same way we will wakeup from deep sleep and will get the RTC alarm correctly fired after
10 seconds.

Signed-off-by: Simon Busch <morphis@gravedo.de>